Implement more feedback of Botho
